Refactor Walkers to make Strategizer walking a special case (subclass) of general walking.
