+++ src/runtime/debug/arc-planner-invoker.ts
@@ -12,7 +12,8 @@ import {Arc} from '../arc.js';
 import {Planner} from '../planner.js';
 import {RecipeIndex} from '../recipe-index.js';
 import {CoalesceRecipes} from '../strategies/coalesce-recipes.js';
-import {Descendant, Strategizer, Strategy, StrategyDerived} from '../../planning/strategizer.js';
+import {Descendant} from '../recipe/walker.js';
+import {Strategizer, Strategy, StrategyDerived} from '../../planning/strategizer.js';
 import * as Rulesets from '../strategies/rulesets.js';
 import {Manifest} from '../manifest.js';
 import {Recipe} from '../recipe/recipe.js';
