+++ src/runtime/strategies/assign-handles.ts
@@ -5,8 +5,7 @@
 // subject to an additional IP rights grant found at
 // http://polymer.github.io/PATENTS.txt
 
-import {Strategizer, Strategy} from '../../planning/strategizer.js';
-import {Walker} from '../recipe/walker.js';
+import {StrategizerWalker, Strategy} from '../../planning/strategizer.js';
 import {Recipe} from '../recipe/recipe.js';
 import {RecipeUtil} from '../recipe/recipe-util.js';
 import {assert} from '../../platform/assert-web.js';
@@ -16,7 +15,7 @@ export class AssignHandles extends Strategy {
   async generate(inputParams) {
     const self = this;
 
-    return Strategizer.over(this.getResults(inputParams), new class extends Walker {
+    return StrategizerWalker.over(this.getResults(inputParams), new class extends StrategizerWalker {
       onHandle(recipe, handle) {
         if (!['?', 'use', 'copy', 'map'].includes(handle.fate)) {
           return undefined;
@@ -90,7 +89,7 @@ export class AssignHandles extends Strategy {
         return score;
       }
 
-    }(Walker.Permuted), this);
+    }(StrategizerWalker.Permuted), this);
   }
 
   getMappableStores(fate, type, tags, counts): Map<StorageProviderBase, string> {
