+++ src/runtime/strategies/convert-constraints-to-connections.ts
@@ -5,9 +5,8 @@
 // subject to an additional IP rights grant found at
 // http://polymer.github.io/PATENTS.txt
 
-import {Strategizer, Strategy} from '../../planning/strategizer.js';
+import {StrategizerWalker, Strategy} from '../../planning/strategizer.js';
 import {Recipe} from '../recipe/recipe.js';
-import {Walker} from '../recipe/walker.js';
 import {RecipeUtil} from '../recipe/recipe-util.js';
 import {ParticleEndPoint, HandleEndPoint, TagEndPoint, InstanceEndPoint} from '../recipe/connection-constraint.js';
 import {Arc} from '../arc.js';
@@ -16,7 +15,7 @@ import {Modality} from '../modality.js';
 export class ConvertConstraintsToConnections extends Strategy {
   async generate(inputParams) {
     const arcModality = this.arc.modality;
-    return Strategizer.over(this.getResults(inputParams), new class extends Walker {
+    return StrategizerWalker.over(this.getResults(inputParams), new class extends StrategizerWalker {
       onRecipe(recipe: Recipe) {
         const modality = arcModality.intersection(recipe.modality);
         // The particles & handles Sets are used as input to RecipeUtil's shape functionality
@@ -220,6 +219,6 @@ export class ConvertConstraintsToConnections extends Strategy {
 
         return processedResults;
       }
-    }(Walker.Independent), this);
+    }(StrategizerWalker.Independent), this);
   }
 }
