+++ src/runtime/strategies/create-description-handle.ts
@@ -5,13 +5,12 @@
 // subject to an additional IP rights grant found at
 // http://polymer.github.io/PATENTS.txt
 
-import {Strategizer, Strategy} from '../../planning/strategizer.js';
+import {StrategizerWalker, Strategy} from '../../planning/strategizer.js';
 import {Recipe} from '../recipe/recipe.js';
-import {Walker} from '../recipe/walker.js';
 
 export class CreateDescriptionHandle extends Strategy {
   async generate(inputParams) {
-    return Strategizer.over(this.getResults(inputParams), new class extends Walker {
+    return StrategizerWalker.over(this.getResults(inputParams), new class extends StrategizerWalker {
       onHandleConnection(recipe, handleConnection) {
         if (handleConnection.handle) {
           return undefined;
@@ -27,6 +26,6 @@ export class CreateDescriptionHandle extends Strategy {
           return 1;
         };
       }
-    }(Walker.Permuted), this);
+    }(StrategizerWalker.Permuted), this);
   }
 }
