+++ src/runtime/strategies/find-hosted-particle.ts
@@ -5,9 +5,8 @@
 // subject to an additional IP rights grant found at
 // http://polymer.github.io/PATENTS.txt
 
-import {Strategizer, Strategy} from '../../planning/strategizer.js';
+import {StrategizerWalker, Strategy} from '../../planning/strategizer.js';
 import {Recipe} from '../recipe/recipe.js';
-import {Walker} from '../recipe/walker.js';
 import {Arc} from '../arc.js';
 import {assert} from '../../platform/assert-web.js';
 import {HandleConnection} from '../recipe/handle-connection.js';
@@ -18,7 +17,7 @@ export class FindHostedParticle extends Strategy {
 
   async generate(inputParams) {
     const arc = this.arc;
-    return Strategizer.over(this.getResults(inputParams), new class extends Walker {
+    return StrategizerWalker.over(this.getResults(inputParams), new class extends StrategizerWalker {
       onHandleConnection(recipe: Recipe, connection: HandleConnection) {
         if (connection.direction !== 'host' || connection.handle) return undefined;
         assert(connection.type instanceof InterfaceType);
@@ -46,6 +45,6 @@ export class FindHostedParticle extends Strategy {
         }
         return results;
       }
-    }(Walker.Permuted), this);
+    }(StrategizerWalker.Permuted), this);
   }
 }
