+++ src/runtime/strategies/init-population.ts
@@ -5,7 +5,8 @@
 // subject to an additional IP rights grant found at
 // http://polymer.github.io/PATENTS.txt
 
-import {Strategy, Descendant} from '../../planning/strategizer.js';
+import {Descendant} from '../recipe/walker.js';
+import {Strategy} from '../../planning/strategizer.js';
 import {Arc} from '../arc.js';
 import {Recipe} from '../recipe/recipe.js';
 import {RecipeIndex} from '../recipe-index.js';
