+++ src/runtime/strategies/match-recipe-by-verb.ts
@@ -5,10 +5,9 @@
 // subject to an additional IP rights grant found at
 // http://polymer.github.io/PATENTS.txt
 
-import {Strategizer, Strategy} from '../../planning/strategizer.js';
+import {StrategizerWalker, Strategy} from '../../planning/strategizer.js';
 import {Particle} from '../recipe/particle.js';
 import {Recipe} from '../recipe/recipe.js';
-import {Walker} from '../recipe/walker.js';
 import {Handle} from '../recipe/handle.js';
 import {Arc} from '../arc.js';
 import {assert} from '../../platform/assert-web.js';
@@ -28,7 +27,7 @@ export class MatchRecipeByVerb extends Strategy {
 
   async generate(inputParams) {
     const arc = this.arc;
-    return Strategizer.over(this.getResults(inputParams), new class extends Walker {
+    return StrategizerWalker.over(this.getResults(inputParams), new class extends StrategizerWalker {
       onParticle(recipe: Recipe, particle: Particle) {
         if (particle.name) {
           // Particle already has explicit name.
@@ -171,7 +170,7 @@ export class MatchRecipeByVerb extends Strategy {
           };
         });
       }
-    }(Walker.Permuted), this);
+    }(StrategizerWalker.Permuted), this);
   }
 
   static satisfiesHandleConstraints(recipe, handleConstraints) {
