+++ src/runtime/strategies/name-unnamed-connections.ts
@@ -5,13 +5,12 @@
 // subject to an additional IP rights grant found at
 // http://polymer.github.io/PATENTS.txt
 
-import {Strategizer, Strategy} from '../../planning/strategizer.js';
+import {StrategizerWalker, Strategy} from '../../planning/strategizer.js';
 import {Recipe} from '../recipe/recipe.js';
-import {Walker} from '../recipe/walker.js';
 
 export class NameUnnamedConnections extends Strategy {
   async generate(inputParams) {
-    return Strategizer.over(this.getResults(inputParams), new class extends Walker {
+    return StrategizerWalker.over(this.getResults(inputParams), new class extends StrategizerWalker {
       onHandleConnection(recipe, handleConnection) {
         if (handleConnection.name) {
           // it is already named.
@@ -32,6 +31,6 @@ export class NameUnnamedConnections extends Strategy {
           };
         });
       }
-    }(Walker.Permuted), this);
+    }(StrategizerWalker.Permuted), this);
   }
 }
