+++ src/runtime/strategies/resolve-recipe.ts
@@ -5,8 +5,7 @@
 // subject to an additional IP rights grant found at
 // http://polymer.github.io/PATENTS.txt
 
-import {Strategizer, Strategy} from '../../planning/strategizer.js';
-import {Walker} from '../recipe/walker.js';
+import {StrategizerWalker, Strategy} from '../../planning/strategizer.js';
 import {Recipe} from '../recipe/recipe.js';
 import {RecipeUtil} from '../recipe/recipe-util.js';
 import {MapSlots} from './map-slots.js';
@@ -19,7 +18,7 @@ export class ResolveRecipe extends Strategy {
 
   async generate(inputParams) {
     const arc = this.arc;
-    return Strategizer.over(this.getResults(inputParams), new class extends Walker {
+    return StrategizerWalker.over(this.getResults(inputParams), new class extends StrategizerWalker {
       onHandle(recipe: Recipe, handle: Handle) {
         if (handle.connections.length === 0 ||
             (handle.id && handle.storageKey) || (!handle.type) ||
@@ -141,6 +140,6 @@ export class ResolveRecipe extends Strategy {
         }
         return undefined;
       }
-    }(Walker.Permuted), this);
+    }(StrategizerWalker.Permuted), this);
   }
 }
