+++ src/runtime/test/strategies/rulesets-tests.ts
@@ -8,14 +8,13 @@
  * http://polymer.github.io/PATENTS.txt
  */
 
-import {Strategizer, Strategy, Ruleset} from '../../../planning/strategizer.js';
+import {StrategizerWalker, Strategy, Ruleset} from '../../../planning/strategizer.js';
 import {Arc} from '../../arc.js';
 import {Manifest} from '../../manifest.js';
 import {Planner} from '../../planner.js';
 import {Recipe} from '../../recipe/recipe.js';
 import {StrategyTestHelper} from './strategy-test-helper.js';
 import {assert} from '../chai-web.js';
-import {Walker} from '../../recipe/walker.js';
 
 class InitPopulation extends Strategy {
   private readonly _context: Manifest;
@@ -50,7 +49,7 @@ class FateAssigner extends Strategy {
 
   async generate(inputParams) {
     const self = this;
-    return Strategizer.over(this.getResults(inputParams), new class extends Walker {
+    return StrategizerWalker.over(this.getResults(inputParams), new class extends StrategizerWalker {
       onHandle(recipe, handle) {
         if (handle.fate === '?') {
           return [
@@ -60,7 +59,7 @@ class FateAssigner extends Strategy {
         }
         return undefined;
       }
-    }(Walker.Permuted), this);
+    }(StrategizerWalker.Permuted), this);
   }
 }
 
@@ -70,7 +69,7 @@ class AssignFateC extends FateAssigner {constructor() {super('use');}}
 
 class Resolve extends Strategy {
   async generate(inputParams) {
-    return Strategizer.over(this.getResults(inputParams), new class extends Walker {
+    return StrategizerWalker.over(this.getResults(inputParams), new class extends StrategizerWalker {
       onHandle(recipe, handle) {
         if (handle.fate !== '?' && !handle.id.endsWith('resolved')) {
           return [
@@ -82,7 +81,7 @@ class Resolve extends Strategy {
         }
         return undefined;
       }
-    }(Walker.Permuted), this);
+    }(StrategizerWalker.Permuted), this);
   }
 }
